# protobuf/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /router/create:
        post:
            tags:
                - RouterHandler
            operationId: RouterHandler_CreateRouter
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.CreateRouterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.CreateRouterResponse'
    /router/delete:
        post:
            tags:
                - RouterHandler
            operationId: RouterHandler_DeleteRouter
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.DeleteRouterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.DeleteRouterResponse'
    /router/info:
        post:
            tags:
                - RouterHandler
            operationId: RouterHandler_GetRouterInfo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.GetRouterInfoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.GetRouterInfoResponse'
    /router/list:
        post:
            tags:
                - RouterHandler
            operationId: RouterHandler_GetRouterList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.GetRouterListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.GetRouterListResponse'
    /router/update:
        post:
            tags:
                - RouterHandler
            operationId: RouterHandler_UpdateRouter
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.UpdateRouterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.UpdateRouterResponse'
    /security/hello:
        post:
            tags:
                - SecurityHandler
            operationId: SecurityHandler_GetPublicKey
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.GetPublicKeyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.GetPublicKeyResponse'
    /security/login:
        post:
            tags:
                - SecurityHandler
            operationId: SecurityHandler_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.LoginResponse'
    /security/logout:
        post:
            tags:
                - SecurityHandler
            operationId: SecurityHandler_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.LogoutResponse'
    /security/router_tree:
        post:
            tags:
                - SecurityHandler
            operationId: SecurityHandler_GetSecurityRouterTree
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.GetSecurityRouterTreeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.GetSecurityRouterTreeResponse'
    /security/self:
        post:
            tags:
                - SecurityHandler
            operationId: SecurityHandler_Self
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.SelfRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.SelfResponse'
    /user/info:
        post:
            tags:
                - UserHandler
            operationId: UserHandler_GetUserInfo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.GetUserInfoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.GetUserInfoResponse'
    /user/list:
        post:
            tags:
                - UserHandler
            operationId: UserHandler_GetUserList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.api.GetUserListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.api.GetUserListResponse'
components:
    schemas:
        server.api.CreateRouterRequest:
            type: object
            properties:
                router:
                    $ref: '#/components/schemas/server.api.CreateRouterRequestData'
        server.api.CreateRouterRequestData:
            type: object
            properties:
                name:
                    type: string
                path:
                    type: string
                title:
                    type: string
                is_menu:
                    type: boolean
                pid:
                    type: string
                sort:
                    type: integer
                    format: int32
                component:
                    type: string
        server.api.CreateRouterResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.CreateRouterResponseData'
        server.api.CreateRouterResponseData:
            type: object
            properties:
                id:
                    type: string
        server.api.DeleteRouterRequest:
            type: object
            properties:
                ids:
                    type: array
                    items:
                        type: string
        server.api.DeleteRouterResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        server.api.GetPublicKeyRequest:
            type: object
            properties: {}
        server.api.GetPublicKeyResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.GetPublicKeyResponseData'
        server.api.GetPublicKeyResponseData:
            type: object
            properties:
                key:
                    type: string
                iv:
                    type: string
        server.api.GetRouterInfoRequest:
            type: object
            properties:
                id:
                    type: string
        server.api.GetRouterInfoResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.Router'
        server.api.GetRouterListRequest:
            type: object
            properties: {}
        server.api.GetRouterListResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.GetRouterListResponseData'
        server.api.GetRouterListResponseData:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/server.api.Router'
        server.api.GetSecurityRouterTreeRequest:
            type: object
            properties: {}
        server.api.GetSecurityRouterTreeResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.GetSecurityRouterTreeResponseData'
        server.api.GetSecurityRouterTreeResponseData:
            type: object
            properties:
                routers:
                    type: array
                    items:
                        $ref: '#/components/schemas/server.api.Router'
        server.api.GetUserInfoRequest:
            type: object
            properties:
                id:
                    type: string
        server.api.GetUserInfoResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.User'
        server.api.GetUserListRequest:
            type: object
            properties:
                currentPage:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                filter:
                    $ref: '#/components/schemas/server.api.GetUserListRequestFilter'
        server.api.GetUserListRequestFilter:
            type: object
            properties: {}
        server.api.GetUserListResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.GetUserListResponseData'
        server.api.GetUserListResponseData:
            type: object
            properties:
                total:
                    type: string
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/server.api.User'
        server.api.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        server.api.LoginResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.LoginResponseData'
        server.api.LoginResponseData:
            type: object
            properties:
                token:
                    type: string
                user:
                    $ref: '#/components/schemas/server.api.User'
        server.api.LogoutRequest:
            type: object
            properties: {}
        server.api.LogoutResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        server.api.Router:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                path:
                    type: string
                title:
                    type: string
                is_menu:
                    type: boolean
                pid:
                    type: string
                sort:
                    type: integer
                    format: int32
                component:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/server.api.Router'
        server.api.SelfRequest:
            type: object
            properties: {}
        server.api.SelfResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/server.api.User'
        server.api.UpdateRouterRequest:
            type: object
            properties:
                router:
                    $ref: '#/components/schemas/server.api.UpdateRouterRequestData'
        server.api.UpdateRouterRequestData:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                path:
                    type: string
                title:
                    type: string
                is_menu:
                    type: boolean
                pid:
                    type: string
                sort:
                    type: integer
                    format: int32
                component:
                    type: string
        server.api.UpdateRouterResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        server.api.User:
            type: object
            properties:
                id:
                    type: string
                username:
                    type: string
tags:
    - name: RouterHandler
    - name: SecurityHandler
    - name: UserHandler